Added HitBox Drawing Added Sword Collision Added Sword Pick up Added Enemy Collision - Needs refining as the enemies get "stuck" on each other

+++ include/sword.h
@@ -21,8 +21,8 @@ public:
 
 	State currentState;
 
-	double damage;
-
+	int damage;
+	
 	Sword(const char*);
 
 	void Init();
